rename child components
remove console log 😬
remove panel if no data
