+++ client/web/src/search/results/components/aggregation/AggregationBarsBackground.module.scss
