+++ client/web/src/search/results/components/aggregation/AggregationDataContainer.module.scss
