+++ client/web/src/search/results/components/aggregation/AggregationErrorContainer.module.scss
